Prevents extra thumbnails in Lockscreen component
Add extra parenthesis when the preview selection is modified to
prevent grouping the component id in the where clause. Without this
patch, multiple "ghost" thumbnails are displayed in the lock screen
selector when a theme includes multiple wallpapers.

This patch will also prevent the WallpaperPreviewGenerator from
overriding the wallpaper thumbnail with the lock screen wallpaper.

Change-Id: I7215a22168073dce99b13a52c2fd780ff6c44a0d
TICKET: CYNGNOS-1758
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>